move endpoint to the snapshot folder
